feat(W42G34): GREEN - deny in-place ledger edits on the Bash channel (#644 G3)
New lib/bash-ledger-writes.ts predicate (sed/perl -i naming a tickets-
namespace test-definitions.md) wired into pre-tool-quality's Bash branch;
shell tokenization extracted from cursor/gate-adapter.ts into
lib/shell-segments.ts so lib never imports from an adapter. Detection
limits documented in the module.
